requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm
scikit-learn
pillow
pyyaml
pandas
timm>=0.9.10
huggingface-hub>=0.23.0
open_clip_torch
numpy
matplotlib
